Clean up 3B2 diagnostic CRC loop warnings
The diagnostic CRC tables in the 3B2 CTC and SCSI code are tiny fixed
arrays. Cast their computed element counts to int32 before comparing
against the existing signed loop index, avoiding -Wsign-compare noise
without changing the loop behavior.

+++ simulators/3B2/3b2_ctc.c
@@ -302,7 +302,7 @@ static void ctc_cmd(uint8 slot,
         /* If the currently running program is a diagnostic program,
          * we are expected to write results into memory at address
          * 0x200f000 */
-        for (i = 0; i < (sizeof(diag_crc) / sizeof(diag_crc[0])); i++) {
+        for (i = 0; i < (int32)(sizeof(diag_crc) / sizeof(diag_crc[0])); i++) {
             if (ctc_crc == diag_crc[i]) {
                 sim_debug(TRACE_DBG, &ctc_dev,
                          "[ctc_cmd] CIO_FCF found CRC==%08x\n", ctc_crc);
